Make current section highlighted instead of previous (#31)
When I go directly to a link, it highlights the link before the
one I am visiting. I put some console.log() statements in and it looks
like it's some sub-pixel rounding problem?

I experience the issue on my own site and also the official tocbot
docs page, in both chromium and firefox (on linux). This patch fixes
the issue for my site in both browsers.